Cleaned SFG a bit
