+++ core/src/main/java/com/volmit/iris/Iris.java
@@ -40,6 +40,7 @@ import com.volmit.iris.engine.platform.BukkitChunkGenerator;
 import com.volmit.iris.engine.platform.DummyChunkGenerator;
 import com.volmit.iris.engine.safeguard.IrisSafeguard;
 import com.volmit.iris.engine.safeguard.ServerBootSFG;
+import com.volmit.iris.engine.safeguard.UtilsSFG;
 import com.volmit.iris.util.collection.KList;
 import com.volmit.iris.util.collection.KMap;
 import com.volmit.iris.util.exceptions.IrisException;
@@ -468,9 +469,9 @@ public class Iris extends VolmitPlugin implements Listener {
             J.s(this::setupPapi);
             J.a(ServerConfigurator::configure, 20);
             splash();
-            ServerBootSFG.UnstableMode();
-            ServerBootSFG.SupportedServerSoftware();
-            ServerBootSFG.printIncompatiblePluginWarnings();
+            UtilsSFG.UnstableMode();
+            UtilsSFG.SupportedServerSoftware();
+            UtilsSFG.printIncompatiblePluginWarnings();
             autoStartStudio();
             checkForBukkitWorlds();
             IrisToolbelt.retainMantleDataForSlice(String.class.getCanonicalName());
